feat: add scoring updates for language, skills, outcomes, and reranker bootstrap (#44)
* feat: add language preference handling in job matching

- Introduced `preferred_language` field in the `profiles` and `jobs` tables.
- Updated `FitScoreComponents` to include `language_match` for scoring.
- Enhanced role catalog with additional search aliases for various roles.
- Implemented skill section parsing to differentiate between required and preferred skills in job descriptions.
- Updated matching logic to account for language preference, affecting overall job fit scores.
- Added tests to validate new language preference scoring and its impact on job matching.

* feat: add application outcome scoring and related functionality

- Introduced `apply_application_outcome_scoring` to adjust job fit scores based on application outcomes.
- Added `outcome_role_boosts` to `SearchLearningAggregates` to store boosts based on application outcomes.
- Implemented `load_application_outcome_role_boosts` to retrieve outcome boosts from the database.
- Created `application_outcome_boost` function to determine boost values based on application status and outcome.
- Updated search ranking logic to incorporate application outcome scoring.
- Added unit tests for new scoring logic and ensured existing tests pass.
- Refactored related code for clarity and maintainability.

* feat: implement job scoring signals and related functionality

* feat: update test assertions for model promotion decision and enhance skill section scoring test

+++ apps/engine-api/src/services/matching/scoring.rs
@@ -151,11 +151,11 @@ pub(super) fn confidence_factor(confidence: Option<u8>) -> f32 {
     }
 }
 
-pub(super) fn weighted_overlap_ratio(matched_strength: f32, total: usize) -> f32 {
-    if total == 0 {
+pub(super) fn weighted_overlap_ratio(matched_strength: f32, total: f32) -> f32 {
+    if total == 0.0 {
         0.0
     } else {
-        (matched_strength / total as f32).min(1.0)
+        (matched_strength / total).min(1.0)
     }
 }
